// apps/nfid-frontend/src/integration/bitcoin/idl/chain-fusion-signer.d.ts
export interface EthSignTransactionRequest {
	to: string;
	gas: bigint;
	value: bigint;
	max_priority_fee_per_gas: bigint;
	data: [] | [string];
	max_fee_per_gas: bigint;
	chain_id: bigint;
	nonce: bigint;
}